interfaces: hide away the IPv6 connectivity flag #7446
I believe the best course of action is to make it the default
and remove the option.

If people want to use the hardware interface and PPPoE at the
same time can simply assign a second interface for the hardware.

This will break fragile IPv6 setups, but is much better in the
long run and will allow further simplification.